Moves/Renames /bot.js to lib/bots/math-bot.js
