+++ lib/bots/math-bot.js
@@ -1,7 +1,7 @@
 'use strict';
 var http = require('http');
-var streams = require('./lib/util/streams');
-var log = require('./lib/util/log');
+var streams = require('./../util/streams');
+var log = require('./../util/log');
 
 var requestOptions = {
   host: 'localhost',
